+++ src/swtpm/Makefile.am
@@ -24,7 +24,8 @@ noinst_HEADERS = \
 	swtpm_aes.h \
 	swtpm_debug.h \
 	swtpm_io.h \
-	swtpm_nvfile.h \
+	swtpm_nvstore.h \
+	swtpm_nvstore_dir.h \
 	threadpool.h \
 	tlv.h \
 	tpmlib.h \
@@ -49,7 +50,8 @@ libswtpm_libtpms_la_SOURCES = \
 	swtpm_aes.c \
 	swtpm_debug.c \
 	swtpm_io.c \
-	swtpm_nvfile.c \
+	swtpm_nvstore.c \
+	swtpm_nvstore_dir.c \
 	tlv.c \
 	tpmlib.c \
 	tpmstate.c \
