+++ src/swtpm/swtpm_nvstore.c
@@ -82,7 +82,8 @@
 #include "swtpm.h"
 #include "swtpm_aes.h"
 #include "swtpm_debug.h"
-#include "swtpm_nvfile.h"
+#include "swtpm_nvstore.h"
+#include "swtpm_nvstore_dir.h"
 #include "key.h"
 #include "logging.h"
 #include "tpmstate.h"
@@ -128,14 +129,9 @@ static encryptionkey migrationkey = {
 
 static uint32_t g_ivec_length;
 static unsigned char *g_ivec;
+static struct nvram_backend_ops *g_nvram_backend_ops;
 
 /* local prototypes */
-
-static TPM_RESULT SWTPM_NVRAM_GetFilenameForName(char *filename,
-                                                 size_t bufsize,
-                                                  uint32_t tpm_number,
-                                                 const char *name,
-                                                 bool is_tempfile);
 
 static TPM_RESULT SWTPM_NVRAM_EncryptData(const encryptionkey *key,
                                           tlv_data *td,
@@ -167,74 +163,6 @@ static TPM_RESULT SWTPM_NVRAM_CheckHeader(unsigned char *data, uint32_t length,
                                           uint8_t *hdrversion,
                                           bool quiet);
 
-/* A file name in NVRAM is composed of 3 parts:
-
-  1 - 'state_directory' is the rooted path to the TPM state home directory
-  2 = 'tpm_number' is the TPM instance, 00 for a single TPM
-  2 - the file name
-
-  For the IBM cryptographic coprocessor version, the root path is hard coded.
-
-  For the Linux and Windows versions, the path comes from an environment variable.  This variable is
-  used once in SWTPM_NVRAM_Init().
-
-  One root path is used for all virtual TPM's, so it can be a static variable.
-*/
-
-static int lockfile_fd = -1;
-
-char state_directory[FILENAME_MAX];
-static TPMLIB_TPMVersion tpmversion = TPMLIB_TPM_VERSION_1_2;
-
-/*
- * SWTPM_NVRAM_Set_TPMVersion()  - set the version of the TPM being used
- */
-void SWTPM_NVRAM_Set_TPMVersion(TPMLIB_TPMVersion version)
-{
-    tpmversion = version;
-}
-
-static TPM_RESULT SWTPM_NVRAM_Lock_Lockfile(const char *directory,
-                                            int *fd)
-{
-    TPM_RESULT rc = 0;
-    char *lockfile = NULL;
-    struct flock flock = {
-        .l_type = F_WRLCK,
-        .l_whence = SEEK_SET,
-        .l_start = 0,
-        .l_len = 0,
-    };
-
-    if (asprintf(&lockfile, "%s/.lock", directory) < 0) {
-        logprintf(STDERR_FILENO,
-                  "SWTPM_NVRAM_Lock_Lockfile: Could not asprintf lock filename\n");
-        return TPM_FAIL;
-    }
-
-    *fd = open(lockfile, O_WRONLY|O_CREAT|O_TRUNC|O_NOFOLLOW, 0660);
-    if (*fd < 0) {
-        logprintf(STDERR_FILENO,
-                  "SWTPM_NVRAM_Lock_Lockfile: Could not open lockfile: %s\n",
-                  strerror(errno));
-        rc = TPM_FAIL;
-        goto exit;
-    }
-
-    if (fcntl(*fd, F_SETLK, &flock) < 0) {
-        logprintf(STDERR_FILENO,
-                  "SWTPM_NVRAM_Lock_Lockfile: Could not lock access to lockfile: %s\n",
-                  strerror(errno));
-        rc = TPM_FAIL;
-        close(*fd);
-        *fd = -1;
-    }
-exit:
-    free(lockfile);
-
-    return rc;
-}
-
 /* SWTPM_NVRAM_Init() is called once at startup.  It does any NVRAM required initialization.
 
    This function sets some static variables that are used by all TPM's.
@@ -242,40 +170,21 @@ exit:
 
 TPM_RESULT SWTPM_NVRAM_Init(void)
 {
+    const char  *backend_uri;
     TPM_RESULT  rc = 0;
-    const char  *tpm_state_path;
-    size_t      length;
-
     TPM_DEBUG(" SWTPM_NVRAM_Init:\n");
 
-    /* TPM_NV_DISK TPM emulation stores in local directory determined by environment variable. */
-    if (rc == 0) {
-        tpm_state_path = tpmstate_get_dir();
-        if (tpm_state_path == NULL) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_Init: Error (fatal), TPM_PATH environment "
-                      "variable not set\n");
-            rc = TPM_FAIL;
-        }
+    backend_uri = tpmstate_get_backend_uri();
+    if (strncmp(backend_uri, "dir://", 6) == 0) {
+        g_nvram_backend_ops = &nvram_dir_ops;
+    } else {
+        logprintf(STDERR_FILENO,
+                  "SWTPM_NVRAM_Init: Unsupported backend.\n");
+        rc = TPM_FAIL;
     }
 
-    /* check that the directory name plus a file name will not overflow FILENAME_MAX */
-    if (rc == 0) {
-        length = strlen(tpm_state_path);
-        if ((length + TPM_FILENAME_MAX) > FILENAME_MAX) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_Init: Error (fatal), TPM state path name "
-                      "%s too large\n", tpm_state_path);
-            rc = TPM_FAIL;
-        }
-    }
-    if (rc == 0) {
-        strcpy(state_directory, tpm_state_path);
-        TPM_DEBUG("SWTPM_NVRAM_Init: Rooted state path %s\n", state_directory);
-    }
-
-    if (rc == 0 && lockfile_fd < 0)
-        rc = SWTPM_NVRAM_Lock_Lockfile(state_directory, &lockfile_fd);
+    if (rc == 0)
+        rc = g_nvram_backend_ops->prepare(backend_uri);
 
     return rc;
 }
@@ -297,101 +206,22 @@ SWTPM_NVRAM_LoadData(unsigned char **data,     /* freed by caller */
                      const char *name)
 {
     TPM_RESULT    rc = 0;
-    size_t        src;
     int           irc;
-    int           fd = -1;
-    char          filename[FILENAME_MAX]; /* rooted file name from name */
     unsigned char *decrypt_data = NULL;
     uint32_t      decrypt_length;
     uint32_t      dataoffset = 0;
     uint8_t       hdrversion = 0;
     uint16_t      hdrflags;
-    struct stat   statbuf;
+    const char    *backend_uri = NULL;
 
     TPM_DEBUG(" SWTPM_NVRAM_LoadData: From file %s\n", name);
     *data = NULL;
     *length = 0;
-    /* open the file */
-    if (rc == 0) {
-        /* map name to the rooted filename */
-        rc = SWTPM_NVRAM_GetFilenameForName(filename, sizeof(filename),
-                                            tpm_number, name, false);
-    }
 
     if (rc == 0) {
-        TPM_DEBUG("  SWTPM_NVRAM_LoadData: Opening file %s\n", filename);
-        fd = open(filename, O_RDONLY);                          /* closed @1 */
-        if (fd < 0) {     /* if failure, determine cause */
-            if (errno == ENOENT) {
-                TPM_DEBUG("SWTPM_NVRAM_LoadData: No such file %s\n",
-                         filename);
-                rc = TPM_RETRY;         /* first time start up */
-            }
-            else {
-                logprintf(STDERR_FILENO,
-                          "SWTPM_NVRAM_LoadData: Error (fatal) opening "
-                          "%s for read, %s\n", filename, strerror(errno));
-                rc = TPM_FAIL;
-            }
-        }
-    }
-
-    if (rc == 0) {
-        if (fchmod(fd, tpmstate_get_mode()) < 0) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_LoadData: Could not fchmod %s : %s\n",
-                      filename, strerror(errno));
-            rc = TPM_FAIL;
-        }
-    }
-
-    /* determine the file length */
-    if (rc == 0) {
-        irc = fstat(fd, &statbuf);
-        if (irc == -1L) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_LoadData: Error (fatal) fstat'ing %s, %s\n",
-                      filename, strerror(errno));
-            rc = TPM_FAIL;
-        }
-    }
-    if (rc == 0) {
-        *length = statbuf.st_size;              /* save the length */
-    }
-    /* allocate a buffer for the actual data */
-    if ((rc == 0) && *length != 0) {
-        TPM_DEBUG(" SWTPM_NVRAM_LoadData: Reading %u bytes of data\n", *length);
-        *data = malloc(*length);
-        if (!*data) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_LoadData: Error (fatal) allocating %u "
-                      "bytes\n", *length);
-            rc = TPM_FAIL;
-        }
-    }
-    /* read the contents of the file into the data buffer */
-    if ((rc == 0) && *length != 0) {
-        src = read(fd, *data, *length);
-        if (src != *length) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_LoadData: Error (fatal), data read of %u "
-                      "only read %lu\n", *length, (unsigned long)src);
-            rc = TPM_FAIL;
-        }
-    }
-    /* close the file */
-    if (fd >= 0) {
-        TPM_DEBUG(" SWTPM_NVRAM_LoadData: Closing file %s\n", filename);
-        irc = close(fd);               /* @1 */
-        if (irc != 0) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_LoadData: Error (fatal) closing file %s\n",
-                      filename);
-            rc = TPM_FAIL;
-        }
-        else {
-            TPM_DEBUG(" SWTPM_NVRAM_LoadData: Closed file %s\n", filename);
-        }
+        backend_uri = tpmstate_get_backend_uri();
+        rc = g_nvram_backend_ops->load(data, length, tpm_number, name,
+                                       backend_uri);
     }
 
     if (rc == 0) {
@@ -455,44 +285,14 @@ SWTPM_NVRAM_StoreData_Intern(const unsigned char *data,
                              TPM_BOOL encrypt         /* encrypt if key is set */)
 {
     TPM_RESULT    rc = 0;
-    uint32_t      lrc;
-    int           irc;
-    int           fd = -1;
-    int           dir_fd = -1;
-    char          tmpfile[FILENAME_MAX];  /* rooted temporary file */
-    char          filename[FILENAME_MAX]; /* rooted file name from name */
     unsigned char *filedata = NULL;
     uint32_t      filedata_length = 0;
     tlv_data      td[3];
     size_t        td_len = 0;
     uint16_t      flags = 0;
+    const char    *backend_uri = NULL;
 
     TPM_DEBUG(" SWTPM_NVRAM_StoreData: To name %s\n", name);
-    if (rc == 0) {
-        /* map name to the rooted filename */
-        rc = SWTPM_NVRAM_GetFilenameForName(filename, sizeof(filename),
-                                            tpm_number, name, false);
-    }
-
-    if (rc == 0) {
-        /* map name to the rooted temporary file */
-        rc = SWTPM_NVRAM_GetFilenameForName(tmpfile, sizeof(tmpfile),
-                                            tpm_number, name, true);
-    }
-
-
-    if (rc == 0) {
-        /* open the file */
-        TPM_DEBUG(" SWTPM_NVRAM_StoreData: Opening file %s\n", tmpfile);
-        fd = open(tmpfile, O_WRONLY|O_CREAT|O_TRUNC|O_NOFOLLOW,
-                  tpmstate_get_mode());                        /* closed @1 */
-        if (fd < 0) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_StoreData: Error (fatal) opening %s for "
-                      "write failed, %s\n", tmpfile, strerror(errno));
-            rc = TPM_FAIL;
-        }
-    }
 
     if (rc == 0) {
         if (encrypt && SWTPM_NVRAM_Has_FileKey()) {
@@ -523,96 +323,10 @@ SWTPM_NVRAM_StoreData_Intern(const unsigned char *data,
     if (rc == 0)
         rc = SWTPM_NVRAM_PrependHeader(&filedata, &filedata_length, flags);
 
-    /* write the data to the file */
     if (rc == 0) {
-        TPM_DEBUG("  SWTPM_NVRAM_StoreData: Writing %u bytes of data\n", length);
-        lrc = write_full(fd, filedata, filedata_length);
-        if (lrc != filedata_length) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_StoreData: Error (fatal), data write "
-                      "of %u only wrote %u\n", filedata_length, lrc);
-            rc = TPM_FAIL;
-        }
-    }
-    if (rc == 0 && fd >= 0) {
-        TPM_DEBUG("  SWTPM_NVRAM_StoreData: Syncing file %s\n", tmpfile);
-        irc = fsync(fd);
-        if (irc != 0) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_StoreData: Error (fatal) syncing file, %s\n",
-                      strerror(errno));
-            rc = TPM_FAIL;
-        } else {
-            TPM_DEBUG("  SWTPM_NVRAM_StoreData: Synced file %s\n", tmpfile);
-        }
-    }
-    if (fd >= 0) {
-        TPM_DEBUG("  SWTPM_NVRAM_StoreData: Closing file %s\n", tmpfile);
-        irc = close(fd);             /* @1 */
-        if (irc != 0) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_StoreData: Error (fatal) closing file\n");
-            rc = TPM_FAIL;
-        }
-        else {
-            TPM_DEBUG("  SWTPM_NVRAM_StoreData: Closed file %s\n", tmpfile);
-        }
-    }
-
-    if (rc == 0 && fd >= 0) {
-        irc = rename(tmpfile, filename);
-        if (irc != 0) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_StoreData: Error (fatal) renaming file: %s\n",
-                      strerror(errno));
-            rc = TPM_FAIL;
-        } else {
-            TPM_DEBUG("  SWTPM_NVRAM_StoreData: Renamed file to %s\n", filename);
-        }
-    }
-
-    /*
-     * Quote from linux man 2 fsync:
-     *  Calling fsync() does not necessarily ensure that the entry in the
-     *  directory containing the file has also reached disk. For that an
-     *  explicit fsync() on a file descriptor for the directory is also needed.
-     */
-    if (rc == 0 && fd >= 0) {
-        TPM_DEBUG(" SWTPM_NVRAM_StoreData: Opening dir %s\n", state_directory);
-        dir_fd = open(state_directory, O_RDONLY);
-        if (dir_fd < 0) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_StoreData: Error (fatal) opening %s for "
-                      "fsync failed, %s\n", state_directory, strerror(errno));
-            rc = TPM_FAIL;
-        }
-    }
-    if (rc == 0 && dir_fd >= 0) {
-        TPM_DEBUG("  SWTPM_NVRAM_StoreData: Syncing dir %s\n", state_directory);
-        irc = fsync(dir_fd);
-        if (irc != 0) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_StoreData: Error (fatal) syncing dir, %s\n",
-                      strerror(errno));
-            rc = TPM_FAIL;
-        } else {
-            TPM_DEBUG("  SWTPM_NVRAM_StoreData: Synced dir %s\n", state_directory);
-        }
-    }
-    if (dir_fd >= 0) {
-        TPM_DEBUG("  SWTPM_NVRAM_StoreData: Closing dir %s\n", state_directory);
-        irc = close(dir_fd);
-        if (irc != 0) {
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_StoreData: Error (fatal) closing dir\n");
-            rc = TPM_FAIL;
-        } else {
-            TPM_DEBUG("  SWTPM_NVRAM_StoreData: Closed dir %s\n", state_directory);
-        }
-    }
-
-    if (rc != 0 && fd >= 0) {
-        unlink(tmpfile);
+        backend_uri = tpmstate_get_backend_uri();
+        rc = g_nvram_backend_ops->store(filedata, filedata_length, tpm_number, name,
+                                        backend_uri);
     }
 
     tlv_data_free(td, td_len);
@@ -631,52 +345,6 @@ TPM_RESULT SWTPM_NVRAM_StoreData(const unsigned char *data,
     return SWTPM_NVRAM_StoreData_Intern(data, length, tpm_number, name, TRUE);
 }
 
-/* SWTPM_NVRAM_GetFilenameForName() constructs a rooted file name from the name.
-
-   The filename is of the form:
-
-   state_directory/tpm_number.name
-
-   A temporary filename used to write to may be created. It shold be rename()'d to
-   the non-temporary filename.
-*/
-
-static TPM_RESULT SWTPM_NVRAM_GetFilenameForName(char *filename,        /* output: rooted filename */
-                                                 size_t bufsize,
-                                                 uint32_t tpm_number,
-                                                 const char *name,      /* input: abstract name */
-                                                 bool is_tempfile)      /* input: is temporary file? */
-{
-    TPM_RESULT res = TPM_SUCCESS;
-    int n;
-    const char *suffix = "";
-
-    TPM_DEBUG(" SWTPM_NVRAM_GetFilenameForName: For name %s\n", name);
-
-    switch (tpmversion) {
-    case TPMLIB_TPM_VERSION_1_2:
-        break;
-    case TPMLIB_TPM_VERSION_2:
-        suffix = "2";
-        break;
-    }
-
-    if (is_tempfile) {
-        n = snprintf(filename, bufsize, "%s/TMP%s-%02lx.%s",
-                     state_directory, suffix, (unsigned long)tpm_number, name);
-    } else {
-        n = snprintf(filename, bufsize, "%s/tpm%s-%02lx.%s",
-                     state_directory, suffix, (unsigned long)tpm_number, name);
-    }
-    if ((size_t)n > bufsize) {
-        res = TPM_FAIL;
-    }
-
-    TPM_DEBUG("  SWTPM_NVRAM_GetFilenameForName: File name %s\n", filename);
-
-    return res;
-}
-
 /* SWTPM_NVRAM_DeleteName() deletes the 'name' from NVRAM
 
    Returns:
@@ -691,26 +359,11 @@ TPM_RESULT SWTPM_NVRAM_DeleteName(uint32_t tpm_number,
                                   const char *name,
                                   TPM_BOOL mustExist)
 {
-    TPM_RESULT  rc = 0;
-    int         irc;
-    char        filename[FILENAME_MAX]; /* rooted file name from name */
+    const char *backend_uri = NULL;
+    backend_uri = tpmstate_get_backend_uri();
 
-    TPM_DEBUG(" SWTPM_NVRAM_DeleteName: Name %s\n", name);
-    /* map name to the rooted filename */
-    rc = SWTPM_NVRAM_GetFilenameForName(filename, sizeof(filename),
-                                        tpm_number, name, false);
-    if (rc == 0) {
-        irc = remove(filename);
-        if ((irc != 0) &&               /* if the remove failed */
-            (mustExist ||               /* if any error is a failure, or */
-             (errno != ENOENT))) {      /* if error other than no such file */
-            logprintf(STDERR_FILENO,
-                      "SWTPM_NVRAM_DeleteName: Error, (fatal) file "
-                      "remove failed, errno %d\n", errno);
-            rc = TPM_FAIL;
-        }
-    }
-    return rc;
+    return g_nvram_backend_ops->delete(tpm_number, name, mustExist,
+                                       backend_uri);
 }
